moves tokens index file into tokens/src
- renames it to `tokens.js`
- re-exports it form the root fo the project via `tokens/index.js`
